requirements.txt
transformers
torch
typer
python-dotenv
huggingface_hub
